% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anno_palette.R
\name{get_anno_palette}
\alias{get_anno_palette}
\title{Generate color palette for plotting at Annogen}
\usage{
get_anno_palette()
}
\value{
Character vector with 16 hexadecimal color codes.
}
\description{
Generate color palette for plotting at Annogen
}
\examples{
annogen_colors = get_palette()
}
