% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anno_palette.R
\name{plot_anno_palette}
\alias{plot_anno_palette}
\title{Plot Annogen color palette}
\usage{
plot_anno_palette()
}
\value{
Plot with Annogen color palette
}
\description{
Plot Annogen color palette
}
\examples{
plot_anno_palette()
}
